fix naming and methods to lesson28addonPoem
