+++ Lesson28AddonPoem/pom.xml
@@ -6,9 +6,9 @@
         <artifactId>c18onlTMS2022</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
-    <artifactId>Lesson28Addon2</artifactId>
+    <artifactId>Lesson28AddonPoem</artifactId>
     <packaging>war</packaging>
-    <name>Lesson28Addon2 Maven Webapp</name>
+    <name>Lesson28AddonPoem Maven Webapp</name>
     <dependencies>
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
@@ -17,6 +17,6 @@
         </dependency>
     </dependencies>
     <build>
-        <finalName>Lesson28Addon2</finalName>
+        <finalName>Lesson28AddonPoem</finalName>
     </build>
 </project>
