+++ Lesson28AddonPoem/src/main/java/by/tms/utils/Constants.java
@@ -3,7 +3,7 @@ package by.tms.utils;
 import lombok.experimental.UtilityClass;
 
 @UtilityClass
-public class Constant {
+public class Constants {
     public static final int ZERO = 0;
-    public static final String PATCH_TO_JSON = "/WEB-INF/classes/poems.json5";
+    public static final String PATCH_TO_JSON = "/WEB-INF/classes/poems.json";
 }
