+++ Lesson28AddonPoem/src/main/java/by/tms/utils/ServletUtils.java
@@ -1,0 +1,17 @@
+package by.tms.utils;
+
+import lombok.experimental.UtilityClass;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.UnsupportedEncodingException;
+
+@UtilityClass
+public class ServletUtils {
+
+    public static void setConfig(HttpServletRequest req, HttpServletResponse resp) throws UnsupportedEncodingException {
+        req.setCharacterEncoding("UTF-8");
+        resp.setCharacterEncoding("UTF-8");
+        resp.setContentType("text/html");
+    }
+}
