+++ Lesson28AddonPoem/src/main/resources/poems/Блішчыць у небе зор пасеў.txt
