+++ Lesson28AddonPoem/src/main/resources/poems/Зіянне месяца.txt
