+++ pom.xml
@@ -41,7 +41,7 @@
         <module>Lesson25and26</module>
         <module>Lesson27and28</module>
         <module>Lesson28Addon</module>
-        <module>Lesson28Addon2</module>
+        <module>Lesson28AddonPoem</module>
     </modules>
 
     <properties>
